requirements.txt
pandas==2.2.3
streamlit==1.41.1
pyyaml==6.0.2
streamlit-extras==0.5.0
duckdb==1.1.3